SysUserController

+++ com.swarm.base/src/main/java/com/swarm/base/dao/SysUserDao.java
@@ -12,6 +12,6 @@ public interface SysUserDao extends JpaRepository<SysUser, Integer> {
 	
 	SysUser findFirstByUsername(String username);
 	
-	long countByUsername(String username);
+	int countByUsername(String username);
 	
 }
